+++ datumaro/plugins/mot_format.py
@@ -13,10 +13,9 @@ import logging as log
 import os
 import os.path as osp
 
-from datumaro.components.extractor import (SourceExtractor,
+from datumaro.components.extractor import (SourceExtractor, Importer,
     DatasetItem, AnnotationType, Bbox, LabelCategories
 )
-from datumaro.components.extractor import Importer
 from datumaro.components.converter import Converter
 from datumaro.util import cast
 from datumaro.util.image import Image
@@ -98,17 +97,7 @@ class MotSeqExtractor(SourceExtractor):
         else:
             raise TypeError("Unexpected type of 'labels' argument: %s" % labels)
         self._categories = self._load_categories(labels)
-        self._items = self._load_items(path)
-
-    def categories(self):
-        return self._categories
-
-    def __iter__(self):
-        for item in self._items.values():
-            yield item
-
-    def __len__(self):
-        return len(self._items)
+        self._items = list(self._load_items(path).values())
 
     @staticmethod
     def _parse_labels(path):
@@ -214,42 +203,10 @@ class MotSeqExtractor(SourceExtractor):
         assert set(seq_info) == {'name', 'imdir', 'framerate', 'seqlength', 'imwidth', 'imheight', 'imext'}, seq_info
 
 class MotSeqImporter(Importer):
-    _EXTRACTOR_NAME = 'mot_seq'
-
     @classmethod
-    def detect(cls, path):
-        return len(cls.find_subsets(path)) != 0
-
-    def __call__(self, path, **extra_params):
-        from datumaro.components.project import Project # cyclic import
-        project = Project()
-
-        subsets = self.find_subsets(path)
-        if len(subsets) == 0:
-            raise Exception("Failed to find 'mot' dataset at '%s'" % path)
-
-        for ann_file in subsets:
-            log.info("Found a dataset at '%s'" % ann_file)
-
-            source_name = osp.splitext(osp.basename(ann_file))[0]
-            project.add_source(source_name, {
-                'url': ann_file,
-                'format': self._EXTRACTOR_NAME,
-                'options': extra_params,
-            })
-
-        return project
-
-    @staticmethod
-    def find_subsets(path):
-        subsets = []
-        if path.endswith('.txt') and osp.isfile(path):
-            subsets = [path]
-        elif osp.isdir(path):
-            p = osp.join(path, 'gt', MotPath.GT_FILENAME)
-            if osp.isfile(p):
-                subsets.append(p)
-        return subsets
+    def find_sources(cls, path):
+        return cls._find_sources_recursive(path, '.txt', 'mot_seq',
+            filename=osp.join('gt', osp.splitext(MotPath.GT_FILENAME)[0]))
 
 class MotSeqGtConverter(Converter):
     DEFAULT_IMAGE_EXT = MotPath.IMAGE_EXT
@@ -310,5 +267,5 @@ class MotSeqGtConverter(Converter):
         labels_file = osp.join(anno_dir, MotPath.LABELS_FILE)
         with open(labels_file, 'w', encoding='utf-8') as f:
             f.write('\n'.join(l.name
-                for l in extractor.categories()[AnnotationType.label].items)
+                for l in extractor.categories()[AnnotationType.label])
             )