+++ tests/test_image_dir_format.py
@@ -2,47 +2,30 @@ import numpy as np
 
 from unittest import TestCase
 
-from datumaro.components.project import Project
-from datumaro.components.extractor import Extractor, DatasetItem
+from datumaro.components.project import Dataset
+from datumaro.components.extractor import DatasetItem
 from datumaro.plugins.image_dir import ImageDirConverter
-from datumaro.util.test_utils import TestDir, compare_datasets
+from datumaro.util.test_utils import TestDir, test_save_and_load
 
 
 class ImageDirFormatTest(TestCase):
     def test_can_load(self):
-        class TestExtractor(Extractor):
-            def __iter__(self):
-                return iter([
-                    DatasetItem(id=1, image=np.ones((10, 6, 3))),
-                    DatasetItem(id=2, image=np.ones((5, 4, 3))),
-                ])
+        dataset = Dataset.from_iterable([
+            DatasetItem(id=1, image=np.ones((10, 6, 3))),
+            DatasetItem(id=2, image=np.ones((5, 4, 3))),
+        ])
 
         with TestDir() as test_dir:
-            source_dataset = TestExtractor()
-
-            ImageDirConverter.convert(source_dataset, save_dir=test_dir)
-
-            project = Project.import_from(test_dir, 'image_dir')
-            parsed_dataset = project.make_dataset()
-
-            compare_datasets(self, source_dataset, parsed_dataset)
+            test_save_and_load(self, dataset, ImageDirConverter.convert,
+                test_dir, importer='image_dir')
 
     def test_relative_paths(self):
-        class TestExtractor(Extractor):
-            def __iter__(self):
-                return iter([
-                    DatasetItem(id='1', image=np.ones((4, 2, 3))),
-                    DatasetItem(id='subdir1/1', image=np.ones((2, 6, 3))),
-                    DatasetItem(id='subdir2/1', image=np.ones((5, 4, 3))),
-                ])
+        dataset = Dataset.from_iterable([
+            DatasetItem(id='1', image=np.ones((4, 2, 3))),
+            DatasetItem(id='subdir1/1', image=np.ones((2, 6, 3))),
+            DatasetItem(id='subdir2/1', image=np.ones((5, 4, 3))),
+        ])
 
         with TestDir() as test_dir:
-            source_dataset = TestExtractor()
-
-            ImageDirConverter.convert(source_dataset, save_dir=test_dir)
-
-            project = Project.import_from(test_dir, 'image_dir')
-            parsed_dataset = project.make_dataset()
-
-            compare_datasets(self, source_dataset, parsed_dataset)
-
+            test_save_and_load(self, dataset, ImageDirConverter.convert,
+                test_dir, importer='image_dir')