+++ tests/test_voc_format.py
@@ -20,7 +20,8 @@ from datumaro.plugins.voc_format.converter import (
 from datumaro.plugins.voc_format.importer import VocImporter
 from datumaro.components.project import Project
 from datumaro.util.image import Image
-from datumaro.util.test_utils import TestDir, compare_datasets
+from datumaro.util.test_utils import (TestDir, compare_datasets,
+    test_save_and_load)
 
 
 class VocFormatTest(TestCase):
@@ -131,16 +132,9 @@ class VocImportTest(TestCase):
 class VocConverterTest(TestCase):
     def _test_save_and_load(self, source_dataset, converter, test_dir,
             target_dataset=None, importer_args=None):
-        converter(source_dataset, test_dir)
-
-        if importer_args is None:
-            importer_args = {}
-        parsed_dataset = VocImporter()(test_dir, **importer_args).make_dataset()
-
-        if target_dataset is None:
-            target_dataset = source_dataset
-
-        compare_datasets(self, expected=target_dataset, actual=parsed_dataset)
+        return test_save_and_load(self, source_dataset, converter, test_dir,
+            importer='voc',
+            target_dataset=target_dataset, importer_args=importer_args)
 
     def test_can_save_voc_cls(self):
         class TestExtractor(TestExtractorBase):