+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -
 
 ### Changed
--
+- Implementation of format plugins simplified (<https://github.com/openvinotoolkit/datumaro/pull/22>)
+- `default` is now a default subset name, instead of `None`. The values are interchangeable. (<https://github.com/openvinotoolkit/datumaro/pull/22>)
+- Improved performance of transforms (<https://github.com/openvinotoolkit/datumaro/pull/22>)
 
 ### Deprecated
 -